FD-46 register with phone number and refactoring

+++ app/products/test_products.py
@@ -7,62 +7,51 @@ import pytest
 async def test_get_products(client, product_factory):
     products = await product_factory.create_batch(3)
 
-    resp = await client.get('api/products/')
+    resp = await client.get("api/products/")
     assert resp.status_code == http.HTTPStatus.OK
     assert len(resp.json()) == len(products)
-    assert sorted([p.id for p in products]) == sorted([data['id'] for data in resp.json()])
+    assert sorted([p.id for p in products]) == sorted([data["id"] for data in resp.json()])
 
-    resp = await client.get(f'api/products/{products[0].id}')
+    resp = await client.get(f"api/products/{products[0].id}")
     assert resp.status_code == http.HTTPStatus.OK
-    assert resp.json()['name'] == products[0].name
+    assert resp.json()["name"] == products[0].name
 
 
 @pytest.mark.asyncio
-async def test_create_product(client, category_factory, token_manager):
+async def test_create_product(client, category_factory, manager_token):
     category = await category_factory()
-    data = {'name': 'product1', 'price': 2.0, 'category_id': f'{category.id}'}
+    data = {"name": "product1", "price": 2.0, "category_id": f"{category.id}"}
 
     resp = await client.post(url="/api/products/", json=data)
     assert resp.status_code == http.HTTPStatus.UNAUTHORIZED
 
-    resp = await client.post(
-        url="/api/products/",
-        json=data,
-        headers={"Authorization": f"Bearer {token_manager}"}
-    )
+    resp = await client.post(url="/api/products/", json=data, headers={"Authorization": f"Bearer {manager_token}"})
     assert resp.status_code == http.HTTPStatus.OK
-    assert resp.json()['name'] == data['name']
+    assert resp.json()["name"] == data["name"]
 
 
 @pytest.mark.asyncio
-async def test_update_product(client, product_factory, token_manager):
+async def test_update_product(client, product_factory, manager_token):
     product = await product_factory()
-    data = {'name': 'product1'}
+    data = {"name": "product1"}
 
-    resp = await client.patch(url=f'/api/products/{product.id}', json=data)
+    resp = await client.patch(url=f"/api/products/{product.id}", json=data)
     assert resp.status_code == http.HTTPStatus.UNAUTHORIZED
 
     resp = await client.patch(
-        url=f'/api/products/{product.id}',
-        json=data,
-        headers={"Authorization": f"Bearer {token_manager}"}
+        url=f"/api/products/{product.id}", json=data, headers={"Authorization": f"Bearer {manager_token}"}
     )
 
     assert resp.status_code == http.HTTPStatus.OK
-    assert resp.json()['name'] == data['name']
+    assert resp.json()["name"] == data["name"]
 
 
 @pytest.mark.asyncio
-async def test_delete_product(client, product_factory, token_manager):
+async def test_delete_product(client, product_factory, manager_token):
     product = await product_factory()
 
-    resp = await client.delete(url=f'/api/products/{product.id}')
+    resp = await client.delete(url=f"/api/products/{product.id}")
     assert resp.status_code == http.HTTPStatus.UNAUTHORIZED
 
-    resp = await client.delete(
-        url=f'/api/products/{product.id}',
-        headers={"Authorization": f"Bearer {token_manager}"}
-    )
+    resp = await client.delete(url=f"/api/products/{product.id}", headers={"Authorization": f"Bearer {manager_token}"})
     assert resp.status_code == http.HTTPStatus.OK
-
-
